requirements.txt
asgiref==3.8.1
Django==5.0.6
djangorestframework==3.15.2
gunicorn==22.0.0
psycopg==3.2.1
psycopg-binary==3.2.1
python-dotenv==1.0.1
sqlparse==0.5.0
typing_extensions==4.12.2
